+++ s3.tf
@@ -9,17 +9,19 @@
 ##
 
 data "template_file" "bucket_policy" {
-  count    = "${length(local.apiary_data_buckets)}"
+  for_each = {
+    for schema in local.schemas_info : "${schema["schema_name"]}" => schema
+  }
   template = "${file("${path.module}/templates/apiary_bucket_policy.json")}"
 
   vars = {
     #if apiary_shared_schemas is empty or contains current schema, allow customer accounts to access this bucket.
-    customer_principal = "${length(var.apiary_shared_schemas) == 0 || contains(var.apiary_shared_schemas, element(concat(local.apiary_managed_schema_names_original, list("")), count.index)) ?
+    customer_principal = "${length(var.apiary_shared_schemas) == 0 || contains(var.apiary_shared_schemas, each.key) ?
       join("\",\"", formatlist("arn:aws:iam::%s:root", var.apiary_customer_accounts)) :
-    "arn:aws:iam::${data.aws_caller_identity.current.account_id}:root"}"
+      "arn:aws:iam::${data.aws_caller_identity.current.account_id}:root"}"
 
-    bucket_name       = "${local.apiary_data_buckets[count.index]}"
-    producer_iamroles = "${replace(lookup(var.apiary_producer_iamroles, element(concat(local.apiary_managed_schema_names_original, list("")), count.index), "arn:aws:iam::${data.aws_caller_identity.current.account_id}:root"), ",", "\",\"")}"
+    bucket_name       = each.value["data_bucket"]
+    producer_iamroles = "${replace(lookup(var.apiary_producer_iamroles, each.key, "arn:aws:iam::${data.aws_caller_identity.current.account_id}:root"), ",", "\",\"")}"
   }
 }
 
@@ -27,20 +29,20 @@ data "template_file" "bucket_policy" {
 ### Apiary S3 data buckets
 ##
 resource "aws_s3_bucket" "apiary_data_bucket" {
-  count         = "${length(local.apiary_data_buckets)}"
-  bucket        = "${element(local.apiary_data_buckets, count.index)}"
+  for_each = {
+    for schema in local.schemas_info : "${schema["schema_name"]}" => schema
+  }
+  bucket        = each.value["data_bucket"]
   acl           = "private"
   request_payer = "BucketOwner"
-  policy        = "${data.template_file.bucket_policy.*.rendered[count.index]}"
-  tags = merge(map("Name", local.apiary_data_buckets[count.index]),
-    var.apiary_tags,
-    jsondecode(lookup(var.apiary_managed_schemas[count.index], "tags", "{}"))
-  )
-
+  policy        = data.template_file.bucket_policy[each.key].rendered
+  tags          = merge(map("Name", each.value["data_bucket"]),
+                        var.apiary_tags,
+                        jsondecode(lookup(each.value, "tags", "{}")))
 
   logging {
     target_bucket = var.apiary_log_bucket == "" ? aws_s3_bucket.apiary_managed_logs_bucket[0].id : var.apiary_log_bucket
-    target_prefix = "${var.apiary_log_prefix}${local.apiary_data_buckets[count.index]}/"
+    target_prefix = "${var.apiary_log_prefix}${each.value["data_bucket"]}/"
   }
 
   lifecycle_rule {
@@ -50,22 +52,24 @@ resource "aws_s3_bucket" "apiary_data_bucket" {
     abort_incomplete_multipart_upload_days = var.s3_lifecycle_abort_incomplete_multipart_upload_days
 
     transition {
-      days          = lookup(var.apiary_managed_schemas[count.index], "s3_lifecycle_policy_transition_period", var.s3_lifecycle_policy_transition_period)
-      storage_class = lookup(var.apiary_managed_schemas[count.index], "s3_storage_class", var.s3_storage_class)
+      days          = lookup(each.value, "s3_lifecycle_policy_transition_period", var.s3_lifecycle_policy_transition_period)
+      storage_class = lookup(each.value, "s3_storage_class", var.s3_storage_class)
     }
 
     dynamic "expiration" {
-      for_each = lookup(var.apiary_managed_schemas[count.index], "s3_object_expiration_days", null) != null ? [1] : []
+      for_each = lookup(each.value, "s3_object_expiration_days", null) != null ? [1] : []
       content {
-        days = lookup(var.apiary_managed_schemas[count.index], "s3_object_expiration_days", null)
+        days = lookup(each.value, "s3_object_expiration_days", null)
       }
     }
   }
 }
 
 resource "aws_s3_bucket_inventory" "apiary_bucket" {
-  count  = var.s3_enable_inventory == true ? "${length(local.apiary_data_buckets)}" : 0
-  bucket = "${aws_s3_bucket.apiary_data_bucket.*.id[count.index]}"
+  for_each = var.s3_enable_inventory == true ? {
+    for schema in local.schemas_info : "${schema["schema_name"]}" => schema
+  } : {}
+  bucket = aws_s3_bucket.apiary_data_bucket[each.key].id
 
   name = local.s3_inventory_prefix
 
@@ -89,8 +93,10 @@ resource "aws_s3_bucket_inventory" "apiary_bucket" {
 }
 
 resource "aws_s3_bucket_public_access_block" "apiary_bucket" {
-  count  = var.s3_block_public_access == true ? "${length(local.apiary_data_buckets)}" : 0
-  bucket = "${aws_s3_bucket.apiary_data_bucket.*.id[count.index]}"
+  for_each = var.s3_block_public_access == true ? {
+    for schema in local.schemas_info : "${schema["schema_name"]}" => schema
+  } : {}
+  bucket = aws_s3_bucket.apiary_data_bucket[each.key].id
 
   block_public_acls   = true
   block_public_policy = true
@@ -98,17 +104,32 @@ resource "aws_s3_bucket_public_access_block" "apiary_bucket" {
 }
 
 resource "aws_s3_bucket_notification" "data_events" {
-  count  = "${var.enable_data_events == "" ? 0 : length(local.apiary_data_buckets)}"
-  bucket = "${aws_s3_bucket.apiary_data_bucket.*.id[count.index]}"
+  for_each = var.enable_data_events == "" ? {} : {
+    for schema in local.schemas_info : "${schema["schema_name"]}" => schema if lookup(schema, "enable_data_events_sqs", "0") == "0"
+  }
+  bucket = aws_s3_bucket.apiary_data_bucket[each.key].id
 
   topic {
-    topic_arn = "${aws_sns_topic.apiary_data_events.*.arn[count.index]}"
+    topic_arn = aws_sns_topic.apiary_data_events[each.key].arn
     events    = ["s3:ObjectCreated:*", "s3:ObjectRemoved:*"]
   }
 }
 
+resource "aws_s3_bucket_notification" "data_queue_events" {
+  for_each = { for schema in local.schemas_info : "${schema["schema_name"]}" => schema if lookup(schema, "enable_data_events_sqs", "0") == "1" }
+  bucket   = aws_s3_bucket.apiary_data_bucket[each.key].id
+
+  queue {
+    queue_arn     = aws_sqs_queue.apiary_data_event_queue[0].arn
+    events        = ["s3:ObjectCreated:*", "s3:ObjectRemoved:*"]
+  }
+}
+
+
 resource "aws_s3_bucket_metric" "paid_metrics" {
-  count  = "${var.enable_s3_paid_metrics == "" ? 0 : length(local.apiary_data_buckets)}"
-  bucket = "${aws_s3_bucket.apiary_data_bucket.*.id[count.index]}"
+  for_each = var.enable_s3_paid_metrics == "" ? {} : {
+    for schema in local.schemas_info : "${schema["schema_name"]}" => schema
+  }
+  bucket = aws_s3_bucket.apiary_data_bucket[each.key].id
   name   = "EntireBucket"
 }